requirements.txt
langchain==0.2.12
langchain_core==0.2.29
langchain_openai==0.1.20
openai==1.40.1
python-dotenv==1.0.1
streamlit==1.37.0
zhipuai==2.1.4.20230806
accelerate
pypdf
modelscope
chromadb